:gem: Add user permissions in components routes

+++ src/components/components.controller.ts
@@ -6,12 +6,23 @@ import {
   Patch,
   Param,
   Delete,
+  UseGuards,
+  NotFoundException,
 } from '@nestjs/common';
 import { ComponentsService } from './components.service';
 import { CreateComponentDto } from './dto/create-component.dto';
 import { UpdateComponentDto } from './dto/update-component.dto';
-import { ApiCreatedResponse, ApiOkResponse, ApiTags } from '@nestjs/swagger';
+import {
+  ApiBearerAuth,
+  ApiCreatedResponse,
+  ApiOkResponse,
+  ApiTags,
+} from '@nestjs/swagger';
 import { ComponentEntity } from './entities/component.entity';
+import { JwtAuthGuard } from 'src/auth/jwt-auth.guard';
+import { UserTypes } from 'src/auth/user-type.decorator';
+import { UserType } from '@prisma/client';
+import { PermissionsGuard } from 'src/auth/permissions.guard';
 
 @ApiTags('Componentes')
 @Controller('api/components')
@@ -19,35 +30,77 @@ export class ComponentsController {
   constructor(private readonly componentsService: ComponentsService) {}
 
   @Post()
+  @UserTypes(UserType.ADMIN)
+  @UseGuards(JwtAuthGuard, PermissionsGuard)
+  @ApiBearerAuth()
   @ApiCreatedResponse({ type: ComponentEntity })
-  create(@Body() createComponentDto: CreateComponentDto) {
-    return this.componentsService.create(createComponentDto);
+  async create(@Body() createComponentDto: CreateComponentDto) {
+    return await this.componentsService.create(createComponentDto);
   }
 
   @Get()
+  @UserTypes(UserType.ADMIN, UserType.PRO, UserType.NORMAL)
+  @UseGuards(JwtAuthGuard, PermissionsGuard)
+  @ApiBearerAuth()
   @ApiOkResponse({ type: ComponentEntity, isArray: true })
-  findAll() {
-    return this.componentsService.findAll();
+  async findAll() {
+    return await this.componentsService.findAll();
+  }
+
+  @Get('category/:category')
+  @UserTypes(UserType.ADMIN, UserType.PRO)
+  @UseGuards(JwtAuthGuard, PermissionsGuard)
+  @ApiBearerAuth()
+  @ApiOkResponse({ type: ComponentEntity })
+  async findByCategory(@Param('category') category: string) {
+    return await this.componentsService.findByCategory(category);
   }
 
   @Get(':id')
+  @UserTypes(UserType.ADMIN, UserType.PRO)
+  @UseGuards(JwtAuthGuard, PermissionsGuard)
+  @ApiBearerAuth()
   @ApiOkResponse({ type: ComponentEntity })
-  findOne(@Param('id') id: string) {
-    return this.componentsService.findOne(id);
+  async findOne(@Param('id') id: string) {
+    const component = await this.componentsService.findOne(id);
+
+    if (!component) {
+      throw new NotFoundException(`Component ${id} does not exist`);
+    }
+
+    return component;
   }
 
   @Patch(':id')
+  @UserTypes(UserType.ADMIN)
+  @UseGuards(JwtAuthGuard, PermissionsGuard)
+  @ApiBearerAuth()
   @ApiOkResponse({ type: ComponentEntity })
-  update(
+  async update(
     @Param('id') id: string,
     @Body() updateComponentDto: UpdateComponentDto,
   ) {
-    return this.componentsService.update(id, updateComponentDto);
+    const component = await this.componentsService.findOne(id);
+
+    if (!component) {
+      throw new NotFoundException(`Component ${id} does not exist`);
+    }
+
+    return await this.componentsService.update(id, updateComponentDto);
   }
 
   @Delete(':id')
+  @UserTypes(UserType.ADMIN)
+  @UseGuards(JwtAuthGuard, PermissionsGuard)
+  @ApiBearerAuth()
   @ApiOkResponse({ type: ComponentEntity })
-  remove(@Param('id') id: string) {
-    return this.componentsService.remove(id);
+  async remove(@Param('id') id: string) {
+    const component = await this.componentsService.findOne(id);
+
+    if (!component) {
+      throw new NotFoundException(`Component ${id} does not exist`);
+    }
+
+    return await this.componentsService.remove(id);
   }
 }
